requirements.txt
colorama>=0.4.6
pwinput>=1.0.3
python-dotenv>=1.0.0
rich>=13.3.5
google-genai